Move types from createLabels to definitions file
