+++ src/utils/definitions.ts
@@ -25,3 +25,60 @@ export type FormData = {
 
 export type HandleChange = (fields: Partial<FormData>) => void;
 export type UpdateWeight = (value: string) => void;
+
+export type Payload = Record<string, string>;
+
+export type Links = {
+  pdf: string;
+  csv: string;
+  zip: string;
+};
+
+export type ParsedResponse = { links: Links; tracking: string[]; labelPrices: string[] };
+
+export type ResponseData = {
+  type: string;
+  message: string;
+  bulkOrder: BulkOrder;
+};
+
+interface BulkOrder {
+  _id: string;
+  user: string;
+  orders: Order[];
+  labelType: string;
+  total: number;
+  status: string;
+  uuid: string;
+  csv: string;
+  pdf: string;
+  zip: string;
+  createdAt: string;
+  updatedAt: string;
+  __v: number;
+  csvLink: string;
+  pdfLink: string;
+  zipLink: string;
+}
+
+interface Order {
+  _id: string;
+  uuid: string;
+  user: string;
+  isApi: boolean;
+  labelType: LabelType;
+  price: number;
+  status: string;
+  pdf: string;
+  tracking: string;
+  tracking_details: unknown[];
+  createdAt: string;
+  updatedAt: string;
+  __v: number;
+}
+
+interface LabelType {
+  _id: string;
+  name: string;
+  uid: string;
+}
